Silence another VS warning
R=titzer@chromium.org
BUG=

Review URL: https://codereview.chromium.org/143973004

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@18712 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/types.h
@@ -355,7 +355,7 @@ struct ZoneTypeConfig {
 
   static int as_bitset(Type* type) {
     ASSERT(is_bitset(type));
-    return reinterpret_cast<intptr_t>(type) >> 1;
+    return static_cast<int>(reinterpret_cast<intptr_t>(type) >> 1);
   }
   static Tagged* as_tagged(Type* type) {
     ASSERT(is_tagged(type));
